src/main/java/iot/empiaurhouse/ambrosia/services/RecipeServiceImpl.java
package iot.empiaurhouse.ambrosia.services;

import iot.empiaurhouse.ambrosia.commandobjects.RecipeCommand;
import iot.empiaurhouse.ambrosia.model.Recipe;
import iot.empiaurhouse.ambrosia.objectconverters.RecipeCommandToRecipe;
import iot.empiaurhouse.ambrosia.objectconverters.RecipeToRecipeCommand;
import iot.empiaurhouse.ambrosia.repositories.RecipeRepository;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;
import java.util.HashSet;
import java.util.Optional;
import java.util.Set;

@Slf4j
@Service
public class RecipeServiceImpl implements RecipeService {

    private final RecipeRepository recipeRepository;
    private final RecipeCommandToRecipe recipeCommandToRecipe;
    private final RecipeToRecipeCommand recipeToRecipeCommand;

    public RecipeServiceImpl(RecipeRepository recipeRepository, RecipeCommandToRecipe recipeCommandToRecipe,RecipeToRecipeCommand recipeToRecipeCommand) {
        this.recipeRepository = recipeRepository;
        this.recipeCommandToRecipe = recipeCommandToRecipe;
        this.recipeToRecipeCommand = recipeToRecipeCommand;
    }

    @Override
    public Set<Recipe> getRecipes() {
        log.debug("Recipe Service is in online!");
        Set<Recipe> recipeSet = new HashSet<>();
        recipeRepository.findAll().iterator().forEachRemaining(recipeSet::add);

        log.debug("Recipe Service is completed!");

        return recipeSet;

    }

    @Override
    public Recipe findById(Long l) {
        Optional<Recipe> recipeOptional = recipeRepository.findById(l);
        if(!recipeOptional.isPresent()){
            throw new RuntimeException("Recipe Not Found!");
        }

        return recipeOptional.get();
    }

    @Override
    @Transactional
    public RecipeCommand saveRecipeCommandObject(RecipeCommand recipeCommand) {
        Recipe stagedRecipe = recipeCommandToRecipe.convert(recipeCommand);
        assert stagedRecipe != null;
        Recipe savedRecipe = recipeRepository.save(stagedRecipe);
        log.debug("Saved " + stagedRecipe.getDescription() + " Recipe Command Object w/ ID: " + savedRecipe.getId() + "!");
        return recipeToRecipeCommand.convert(savedRecipe);
    }
}
